third_party/WebKit/LayoutTests/fast/forms/radio/state-restore-radio-group-expected.txt

Test to restore form value states for a radio group.

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("input1").checked is true
PASS document.getElementById("input2").checked is false

 
